refactor: reorganize epics
