+++ src/epics/game_general/nextRoundEpic.ts
@@ -6,8 +6,8 @@ import {
   RESOURCE_PROD,
   SWITCH_NEW_TURN,
   ABORT_ALL,
-} from '../constants/ActionTypes'
-import { RootActionType } from '../types/actionObj'
+} from '../../constants/ActionTypes'
+import { RootActionType } from '../../types/actionObj'
 import {
   withLatestFrom,
   filter,
@@ -17,7 +17,7 @@ import {
 } from 'rxjs/operators'
 import { isOfType } from 'typesafe-actions'
 import { ActionsObservable, StateObservable } from 'redux-observable'
-import { RootStateType } from '../types/state'
+import { RootStateType } from '../../types/state'
 import { concat, of } from 'rxjs'
 
 export default (
