+++ src/epics/index.ts
@@ -2,30 +2,30 @@ import { RootActionType } from '../types/actionObj'
 import { combineEpics, Epic } from 'redux-observable'
 import { RootStateType } from '../types/state'
 
-import updateLangEpic from './updateLangEpic'
-import updateErathianEpic from './updateErathianEpic'
-import updateVolumeEpic from './updateVolumeEpic'
-import updateSettingsEpic from './updateSettingsEpic'
-import updateStatusEpic from './updateStatusEpic'
-import execCardEpic from './execCardEpic'
-import useCardEpic from './useCardEpic'
-import discardCardEpic from './discardCardEpic'
-import moveCardToTopEpic from './moveCardToTopEpic'
-import changeSettingsAndInitEpic from './changeSettingsAndInitEpic'
-import readlsUpdatestoreInitEpic from './readlsUpdatestoreInitEpic'
-import initEpic from './initEpic'
-import drawCardEpic from './drawCardEpic'
-import nextRoundEpic from './nextRoundEpic'
-import resourceProdEpic from './resourceProdEpic'
-import clearCardEpic from './clearCardEpic'
-import checkUnusableEpic from './checkUnusableEpic'
-import checkVictoryEpic from './checkVictoryEpic'
-import screenEndEpic from './screenEndEpic'
-import aiUseCardEpic from './aiUseCardEpic'
-import checkSurrenderEpic from './checkSurrenderEpic'
-import switchMultiplayerModeEpic from './switchMultiplayerModeEpic'
-import getAndSetYourIdEpic from './getAndSetYourIdEpic'
-import connectToAndSetOpponentIdEpic from './connectToAndSetOpponentIdEpic'
+import updateLangEpic from './settings_lang_etc/updateLangEpic'
+import updateErathianEpic from './settings_lang_etc/updateErathianEpic'
+import updateVolumeEpic from './settings_lang_etc/updateVolumeEpic'
+import updateSettingsEpic from './settings_lang_etc/updateSettingsEpic'
+import updateStatusEpic from './status/updateStatusEpic'
+import execCardEpic from './cards/execCardEpic'
+import useCardEpic from './cards/useCardEpic'
+import discardCardEpic from './cards/discardCardEpic'
+import moveCardToTopEpic from './cards/moveCardToTopEpic'
+import changeSettingsAndInitEpic from './settings_lang_etc/changeSettingsAndInitEpic'
+import readlsUpdatestoreInitEpic from './settings_lang_etc/readlsUpdatestoreInitEpic'
+import initEpic from './game_general/initEpic'
+import drawCardEpic from './cards/drawCardEpic'
+import nextRoundEpic from './game_general/nextRoundEpic'
+import resourceProdEpic from './status/resourceProdEpic'
+import clearCardEpic from './cards/clearCardEpic'
+import checkUnusableEpic from './cards/checkUnusableEpic'
+import checkVictoryEpic from './game_general/checkVictoryEpic'
+import screenEndEpic from './screen/screenEndEpic'
+import aiUseCardEpic from './cards/aiUseCardEpic'
+import checkSurrenderEpic from './game_general/checkSurrenderEpic'
+import switchMultiplayerModeEpic from './multiplayer/switchMultiplayerModeEpic'
+import getAndSetYourIdEpic from './multiplayer/getAndSetYourIdEpic'
+import connectToAndSetOpponentIdEpic from './multiplayer/connectToAndSetOpponentIdEpic'
 
 export type MyEpic = Epic<RootActionType, RootActionType, RootStateType>
