+++ src/epics/multiplayer/getAndSetYourIdEpic.ts
@@ -2,8 +2,8 @@ import {
   GET_AND_SET_YOUR_ID,
   SET_YOUR_ID,
   ABORT_ALL,
-} from '../constants/ActionTypes'
-import { RootActionType } from '../types/actionObj'
+} from '../../constants/ActionTypes'
+import { RootActionType } from '../../types/actionObj'
 import {
   withLatestFrom,
   filter,
@@ -14,7 +14,7 @@ import {
 import { of, concat } from 'rxjs'
 import { isOfType } from 'typesafe-actions'
 import { ActionsObservable, StateObservable } from 'redux-observable'
-import { RootStateType } from '../types/state'
+import { RootStateType } from '../../types/state'
 
 export default (
   action$: ActionsObservable<RootActionType>,
