+++ src/epics/multiplayer/switchMultiplayerModeEpic.ts
@@ -1,9 +1,9 @@
 import {
-  CONNECT_TO_AND_SET_OPPONENT_ID,
-  SET_OPPONENT_ID,
+  SWITCH_MULTIPLAYER_MODE,
+  SWITCH_MULTIPLAYER_MODE_MAIN,
   ABORT_ALL,
-} from '../constants/ActionTypes'
-import { RootActionType } from '../types/actionObj'
+} from '../../constants/ActionTypes'
+import { RootActionType } from '../../types/actionObj'
 import {
   withLatestFrom,
   filter,
@@ -14,21 +14,21 @@ import {
 import { of, concat } from 'rxjs'
 import { isOfType } from 'typesafe-actions'
 import { ActionsObservable, StateObservable } from 'redux-observable'
-import { RootStateType } from '../types/state'
+import { RootStateType } from '../../types/state'
 
 export default (
   action$: ActionsObservable<RootActionType>,
   state$: StateObservable<RootStateType>,
 ) =>
   action$.pipe(
-    filter(isOfType(CONNECT_TO_AND_SET_OPPONENT_ID)),
+    filter(isOfType(SWITCH_MULTIPLAYER_MODE)),
     withLatestFrom(state$),
     concatMap(([action, state]) => {
-      const { id } = action
+      const { on } = action
       return concat(
         of<RootActionType>({
-          type: SET_OPPONENT_ID,
-          id,
+          type: SWITCH_MULTIPLAYER_MODE_MAIN,
+          on,
         }),
       ).pipe(takeUntil(action$.ofType(ABORT_ALL)))
     }),
