+++ src/epics/settings_lang_etc/changeSettingsAndInitEpic.ts
@@ -3,13 +3,13 @@ import {
   UPDATE_SETTINGS,
   INIT,
   ABORT_ALL,
-} from '../constants/ActionTypes'
-import { RootActionType } from '../types/actionObj'
+} from '../../constants/ActionTypes'
+import { RootActionType } from '../../types/actionObj'
 import { filter, concatMap, delay } from 'rxjs/operators'
 import { of, concat } from 'rxjs'
 import { isOfType } from 'typesafe-actions'
 import { ActionsObservable, StateObservable } from 'redux-observable'
-import { RootStateType } from '../types/state'
+import { RootStateType } from '../../types/state'
 
 export default (
   action$: ActionsObservable<RootActionType>,
