+++ src/epics/settings_lang_etc/updateLangEpic.ts
@@ -1,36 +1,36 @@
 import {
-  UPDATE_ERATHIAN,
-  UPDATE_ERATHIAN_MAIN,
+  UPDATE_LANG,
+  UPDATE_LANG_MAIN,
   ABORT_ALL,
-} from '../constants/ActionTypes'
-import { RootActionType } from '../types/actionObj'
+} from '../../constants/ActionTypes'
+import { RootActionType } from '../../types/actionObj'
 import { filter, concatMap, takeUntil } from 'rxjs/operators'
 import { of, concat } from 'rxjs'
 import { isOfType } from 'typesafe-actions'
 import { ActionsObservable, StateObservable } from 'redux-observable'
-import { RootStateType } from '../types/state'
-import { lsSet } from '../utils/localstorage'
-import { defaultLang } from '../i18n/langs'
+import { RootStateType } from '../../types/state'
+import { lsSet } from '../../utils/localstorage'
+import { defaultErathian } from '../../constants/defaultSettings'
 
 export default (
   action$: ActionsObservable<RootActionType>,
   state$: StateObservable<RootStateType>,
 ) =>
   action$.pipe(
-    filter(isOfType(UPDATE_ERATHIAN)),
+    filter(isOfType(UPDATE_LANG)),
     concatMap((action) => {
-      const { erathian } = action
+      const { lang } = action
       lsSet((draft) => {
         if (draft.lang === undefined) {
-          draft.lang = { code: defaultLang, erathian }
+          draft.lang = { code: lang, erathian: defaultErathian }
         } else {
-          draft.lang.erathian = erathian
+          draft.lang.code = lang
         }
       })
       return concat(
         of<RootActionType>({
-          type: UPDATE_ERATHIAN_MAIN,
-          erathian,
+          type: UPDATE_LANG_MAIN,
+          lang,
         }),
       ).pipe(takeUntil(action$.ofType(ABORT_ALL)))
     }),
